requirements.txt
alabaster==1.0.0
anaconda-anon-usage @ file:///croot/anaconda-anon-usage_1749054771823/work
annotated-doc==0.0.2
annotated-types @ file:///work/perseverance-python-buildout/croot/annotated-types_1728386958747/work
anyio==4.11.0
appnope==0.1.4
archspec @ file:///croot/archspec_1709217642129/work
asgiref==3.9.2
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
backcall==0.2.0
bcrypt==5.0.0
beautifulsoup4==4.13.5
bleach==6.2.0
boltons @ file:///croot/boltons_1751383775419/work
brotlicffi @ file:///croot/brotlicffi_1736182461069/work
certifi @ file:///croot/certifi_1752653701669/work/certifi
cffi @ file:///croot/cffi_1736182485317/work
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.2.1
comm==0.2.3
conda @ file:///croot/conda_1749171252582/work/conda-src
conda-anaconda-telemetry @ file:///croot/conda-anaconda-telemetry_1751061994888/work
conda-anaconda-tos @ file:///croot/conda-anaconda-tos_1752801564376/work
conda-content-trust @ file:///work/perseverance-python-buildout/croot/conda-content-trust_1728487058234/work
conda-libmamba-solver @ file:///croot/conda-libmamba-solver_1745607008911/work/src
conda-package-handling @ file:///croot/conda-package-handling_1731369017509/work
conda_package_streaming @ file:///croot/conda-package-streaming_1751307506088/work
contourpy==1.3.2
cryptography @ file:///croot/cryptography_1748977429425/work
cycler==0.12.1
decorator==5.2.1
deep-sort-realtime==1.3.2
defusedxml==0.7.1
dill==0.4.0
distro @ file:///work/perseverance-python-buildout/croot/distro_1728396110052/work
dj-database-url==3.0.1
Django==5.2.6
django-cors-headers==4.9.0
django-spa==0.3.6
djangorestframework==3.16.1
dnspython==2.8.0
docutils==0.21.2
ecdsa==0.19.1
email-validator==2.3.0
executing==2.2.1
fastapi==0.120.0
fastapi-cli==0.0.14
fastapi-cloud-cli==0.3.1
fastjsonschema==2.21.2
fica==0.4.1
filelock==3.18.0
filterpy==1.4.5
fonttools==4.59.0
frozendict @ file:///work/perseverance-python-buildout/croot/frozendict_1728497542215/work
fsspec==2025.7.0
git-filter-repo==2.47.0
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna @ file:///work/perseverance-python-buildout/croot/idna_1728385935861/work
imagesize==1.4.1
ipylab==1.1.0
ipython==9.5.0
ipython-genutils==0.2.0
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
jedi==0.19.2
Jinja2==3.1.6
jsonpatch @ file:///work/perseverance-python-buildout/croot/jsonpatch_1728399595941/work
jsonpointer==2.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
jupyterlab_pygments==0.3.0
jupyterlab_widgets==3.0.15
jupytext==1.17.3
kiwisolver==1.4.8
lap==0.5.12
libmambapy @ file:///croot/mamba-split_1734469461757/work/libmambapy
mangum==0.19.0
markdown-it-py @ file:///work/perseverance-python-buildout/croot/markdown-it-py_1728387994477/work
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdit-py-plugins==0.5.0
mdurl @ file:///work/perseverance-python-buildout/croot/mdurl_1728387286143/work
menuinst @ file:///croot/menuinst_1750861706497/work
mistune==3.1.4
mpmath==1.3.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
otter-grader==6.1.4
packaging @ file:///croot/packaging_1734472117206/work
pandas==2.3.1
pandocfilters==1.5.1
parso==0.8.5
passlib==1.7.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs @ file:///croot/platformdirs_1744273042065/work
playwright==1.55.0
pluggy @ file:///croot/pluggy_1733169602837/work
prompt_toolkit==3.0.52
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyasn1==0.6.1
pycosat @ file:///croot/pycosat_1736868416091/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic @ file:///croot/pydantic_1750768261732/work
pydantic_core @ file:///croot/pydantic-core_1750754736071/work
pyee==13.0.0
Pygments @ file:///croot/pygments_1744664109463/work
pyparsing==3.2.3
PySocks @ file:///work/perseverance-python-buildout/croot/pysocks_1728386193338/work
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.5.0
python-multipart==0.0.20
python-on-whales==0.78.0
pytz==2025.2
PyYAML==6.0.2
pyzmq==27.1.0
redis==7.0.0
referencing==0.36.2
requests @ file:///croot/requests_1750426329888/work
rich @ file:///croot/rich_1732638981168/work
rich-toolkit==0.15.1
rignore==0.7.1
roman-numerals-py==3.1.0
rpds-py==0.27.1
rsa==4.9.1
ruamel.yaml @ file:///croot/ruamel.yaml_1745960305322/work
ruamel.yaml.clib @ file:///croot/ruamel.yaml.clib_1745937152469/work
scipy==1.16.0
seaborn==0.13.2
sentry-sdk==2.42.1
setuptools==78.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
soupsieve==2.8
Sphinx==8.2.3
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
tinycss2==1.4.0
toml==0.10.2
torch==2.7.1
torchvision==0.22.1
tornado==6.5.2
tqdm @ file:///croot/tqdm_1738943501192/work
traitlets==5.14.3
triton==3.3.1
truststore @ file:///croot/truststore_1736550121485/work
typer==0.20.0
typing-inspection @ file:///croot/typing-inspection_1746023470701/work
typing_extensions @ file:///croot/typing_extensions_1734714854207/work
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3 @ file:///croot/urllib3_1750775463400/work
uvicorn==0.38.0
uvloop==0.22.1
watchfiles==1.1.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
wheel==0.45.1
whitenoise==5.3.0
widgetsnbextension==4.0.14
wrapt==1.17.3
zstandard @ file:///croot/zstandard_1731356346222/work
